Add files via upload
I've created the collection view with cells that have an image and description for each activity. To add a cell to the collection you simply have to add an activity to the activities array. I also created labels that can show the date and time on the home screen. Let me know what you guys think and I'll start looking into making the cells navigate to the respective activity's instructions when touched.